Split the handler into its own module.
